Extracted duplicated code to base.js
This code defines the editor used in program, function, and debugging
questions. Now instead of being repeated exactly in each, it is defined
once in base.js
